Update README.md
Grading

Step 1

Run the program three times using 80, 55 and 60 for the average. Notice that nothing happens when you input 60 as the average. Modify the first if statement so that the program will also print “You Pass” if the average equals 60.

Step 2

Modify the program so that it uses an if/else statement rather than two if statements.

Step 3

Modify the program from Exercise 2 to allow the following categories:

Invalid data (data above 100)
"A" category (90–100)
"B" category (80–89)
"You Pass" category (60–79)
"You Fail" category (59 or less)
Use Lab 9.2 as an example for using multiway if-else statements for testing a series of conditions to find what range a number is within.